% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_dat.R
\docType{data}
\name{cl_dat}
\alias{cl_dat}
\title{The example of dycutdf output}
\format{An object of list with 10 length.}
\usage{
data(cl_dat)
}
\description{
The list of data frame which result of 'dycutdf.R'.
    These of dataframe objects cutting of hieralchicalclustering.
}
\examples{
data(cl_dat)
lapply(cl_dat, dim)
lapply(cl_dat, function(x){x[1:3,1:3]})

# search target genes conataining cluster ----
sct <- c("gene1", "gene2")
sapply(cl_dat, function(x)match(sct, names(x)))
}
\keyword{datasets}
