% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_dat.R
\docType{data}
\name{cluster_dat}
\alias{cluster_dat}
\title{The example of cluster_mat output}
\format{An object of list with 4 length.}
\usage{
data(cluster_dat)
}
\description{
The list of data frame which result of 'cluster_mat.R'.
    These of dataframe objects cutting of hieralchicalclustering.
}
\examples{
data(cluster_dat)
lapply(cluster_dat, dim)
lapply(cluster_dat, function(x){x[1:6,1:6]})


}
\keyword{datasets}
