% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coredge.R
\name{coredge}
\alias{coredge}
\title{Construction of a network graph from correlation matxix}
\usage{
coredge(mat, it = seq(0.30,0.99,by=0.01))
}
\arguments{
\item{mat}{correlation matrix}

\item{it}{a vector of threshold sequence as iteration. The initial value as default, 'it =seq(0.30,0.99,by=0.01)'}
}
\value{
A list consists from one igraph object and two dataframes. The one is a weighted edge list, the another one is
    a result from ks-test of difference eigen sequence based on thresh correlation matrix.
}
\description{
Construction of a correlation network graph from correlation matrix through Random Matrix Theory-based methods.
}
\examples{
# sample data
data(cluster_dat)
cormat <- cor(cluster_dat[[2]])
res <- coredge(mat = cormat)
res[[1]]
head(res[[2]])
head(res[[3]])
}
