% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corheat.R
\name{corheat}
\alias{corheat}
\title{Gene clustering and cutting cluster with dynamic cut}
\usage{
corheat(dat, distm, clm, method_dycut, draw)
}
\arguments{
\item{dat}{RNA-seq count data(row:sample, column:gene)}

\item{distm}{select from c("pearson", "abspearson", "squarepearson", "spearman")}

\item{clm}{clustering method of hclust}

\item{method_dycut}{method of 'cutreeDynamic' select from "tree", or "hybrid}

\item{draw}{logical draw heatmap or nod.}
}
\value{
draw heatmap and results of dynamic cut
}
\description{
Gene clustering for RNA-seq data and cutting cluster with dynamic cut
}
\examples{
# data: normalized fpkm
fp <- system.file("extdata/nfpkm_rnsq.txt", package = "rsko")
nfpkm <- read.table(fp, header=TRUE, stringsAsFactors = FALSE)

# remove row count
dat <- nfpkm[-1:-3][colSums(nfpkm[-1:-3]) != 0]

# corheat with dynamic tree cut
res <- corheat(dat=dat, distm="spearman", clm="average", method_dycut="tree", draw=TRUE)
}
