% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgelist.R
\name{correlation_edgelist}
\alias{correlation_edgelist}
\title{Correlation edgelist}
\usage{
correlation_edgelist(data, correlation_method = c("pearson", "kendall",
  "spearman"), use = c("everything", "all.obs", "complete.obs",
  "na.or.complete", "pairwise.complete.obs"), threshold = 0.05,
  threshold_method = c("pvalue", "absolute"),
  adjust = stats::p.adjust.methods)
}
\arguments{
\item{data}{input data, a numeric matrix or data frame.}

\item{correlation_method}{a character string indicating which correlatin coefficient to compute. 
One of `"pearson"` (default), `"kendall"`, or `"spearman"`.}

\item{use}{a character string indicating how to deal with missing values. 
One of `"everything"`, `"all.obs"`, `"complete.obs"`, `"na.or.complete"`, or `"pairwise.complete.obs"`.}

\item{threshold}{a numeric threshold}

\item{threshold_method}{a character string indicating how correlations should be thresholded. 
One of `"pvalue"` (default), or `"absolute"`.}

\item{adjust}{a character string indicating how p-values should be adjusted for multiple comparisons. 
See \link[stats]{p.adjust.methods} for all methods.}
}
\description{
Calculate the edgelist of a correlation network
}
\examples{
data('mtcars')
correlation_edgelist(mtcars)

}
