% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{correlation_network}
\alias{correlation_network}
\title{Correlation network}
\usage{
correlation_network(data, node_data = tibble::tibble(name =
  colnames(data)), correlation_method = c("pearson", "kendall",
  "spearman"), use = c("everything", "all.obs", "complete.obs",
  "na.or.complete", "pairwise.complete.obs"), threshold = 0.05,
  threshold_method = c("pvalue", "absolute"),
  adjust = stats::p.adjust.methods)
}
\arguments{
\item{data}{input data, a numeric matrix or data frame.}

\item{node_data}{additional data frame containing information for each node. 
Should have a column `'name'` that contains the column names of the data.}

\item{correlation_method}{a character string indicating which correlatin coefficient to compute. 
One of `"pearson"` (default), `"kendall"`, or `"spearman"`.}

\item{use}{a character string indicating how to deal with missing values. 
One of `"everything"`, `"all.obs"`, `"complete.obs"`, `"na.or.complete"`, or `"pairwise.complete.obs"`.}

\item{threshold}{a numeric threshold}

\item{threshold_method}{a character string indicating how correlations should be thresholded. 
One of `"pvalue"` (default), or `"absolute"`.}

\item{adjust}{a character string indicating how p-values should be adjusted for multiple comparisons. 
See \link[stats]{p.adjust.methods} for all methods.}
}
\description{
Correlation network as a tidy graph
}
\examples{
data('mtcars')
graph_cors <- correlation_network(mtcars)
graph_cors
\dontrun{
require(ggraph)
set.seed(42)
ggraph(graph_cors, layout = "graphopt") + 
  geom_edge_link(aes(color = r), width = 1.3) +
  geom_node_label(aes(label = name)) + 
  scale_edge_colour_gradient2(limits = c(-1, 1)) +
  theme_graph()
}

}
