% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dycutdf.R
\name{dycutdf}
\alias{dycutdf}
\title{Hierarchical clustering using amp, and split dataframe by cluster, with dynamic cut.}
\usage{
dycutdf(dat, distm, clm, column, method_dycut, ...)
}
\arguments{
\item{dat}{data frame or matrix}

\item{distm}{distance measure from amap::Dist options. default is "pearson"}

\item{clm}{clustering method select from hclust options default is "average"}

\item{column}{column which containing expression data default 'column=1:ncol(dat)'}

\item{method_dycut}{method of dynamic cut, otherwise k number of cuttree function.}

\item{...}{additional dynamicTreeCut::cutreeDynamic options}
}
\value{
list of results of clustering, objects was hclust, dendrogram, and list of splitted data.frame
}
\description{
dycutdf
}
\examples{
d <- data.frame(t(iris[-5]))
res <- dycutdf(dat = d, distm = "euclidean", clm = "average",
               column = 1:150, method_dycut = "hybrid")
res[[1]] # hclust object
res[[2]] # result of dynamic cut object
res[[3]] # list of splitted data.frame

##  The following examples takes too much times
# fp <- system.file("extdata/nfpkm_200.txt", package = "cornet")
# nfpkm <- read.table(fp, header=TRUE, stringsAsFactors = FALSE)
# res1 <- cornet::dycutdf(dat = nfpkm, distm = "abscorrelation", clm = "average",
#     column = 5:ncol(nfpkm), method_dycut = "tree")
}
