% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gethub.R
\name{gethub}
\alias{gethub}
\title{node centrality}
\usage{
gethub(g, com_fun)
}
\arguments{
\item{g}{igraph object}

\item{com_fun}{character: function name of igraph for community detection}
}
\description{
hub do.
}
\examples{
##
}
