% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_to_edge.R
\name{mat_to_edge}
\alias{mat_to_edge}
\title{Convert a correlation matrix to a weighted edge list}
\usage{
mat_to_edge(mat)
}
\arguments{
\item{mat}{A squared matrix, and lower and upper triangled data to be a same.}
}
\description{
Convert a correlation matrix to a weighted edge list.
    if this matrix is named matrix, these names as node id. Only lower triangled data is used.
}
\examples{
mat <- cor(iris[-5])
mat_to_edge(mat)
}
