% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matoedge.R
\name{matoedge}
\alias{matoedge}
\title{Convert a correlation matrix to a weighted edge list}
\usage{
matoedge(mat)
}
\arguments{
\item{mat}{A squared matrix, and lower and upper triangled data to be a same.}
}
\description{
Convert a correlation matrix to a weighted edge list.
    if this matrix is named matrix, these names as node id. Only lower triangled data is used.
}
\examples{
mat <- cor(iris[-5])
matoedge(mat)
}
