% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_mat.R
\name{cluster_mat}
\alias{cluster_mat}
\title{matplot per cluster}
\usage{
cluster_mat(dat, res_dycut, fcdat)
}
\arguments{
\item{dat}{a data frame of count data for transcriptome, which columns are genes and row are samples.}

\item{res_dycut}{result of "dynamicTreeCut::cutreeDynamic", or "cutree"}

\item{fcdat}{data.frame of one or two factor for samples, which first column is a factor for Y-axis and a second column is a factor for X-axis}
}
\description{
split count data of transcriptome with result of gene clustering, such as "cutree" or "cutreeDynamic".
}
\examples{
# # argument1: data
# nfpkm <- rskodat::nfpkm

# # argument2: result of dycutdf
# res_dycut <- dycutdf(nfpkm, distm = "abscorrelation", column = -1:-4)

# # argument3: factorial data.frame one(x) or two(y,x)
# fcdat <- data.frame(runs=nfpkm$runs, days=nfpkm$days)
# cluster_mat(nfpkm[-1:-4], res_dycut$dynamic_cut, fcdat)

}
