% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_mat.R
\name{cluster_mat}
\alias{cluster_mat}
\title{Cluster_mat}
\usage{
cluster_mat(dat, distm, clm, column, method_dycut, y_fctr, x_fctr, rep_fctr, ...)
}
\arguments{
\item{dat}{data frame or matrix}

\item{distm}{distance measure from amap::Dist}

\item{clm}{hclust methods}

\item{column}{column which containing expression data default 'column=1:ncol(dat)'}

\item{method_dycut}{method of dynamic cut}

\item{y_fctr}{optional}

\item{x_fctr}{x-axis factor for ggplot object like matplot}

\item{rep_fctr}{optional}

\item{...}{additional dynamicTreeCut::cutreeDynamic options}
}
\description{
Cluster_mat
}
\examples{
d <- data.frame(t(iris[-5]))
xfctr <- factor(names(iris[-5]), levels = names(iris[-5]))
res <- cluster_mat(dat = d, distm = "euclidean", clm = "average",
       column = 1:150, method_dycut = "hybrid", x_fctr = xfctr)
res[[1]] # hclust object
res[[2]] # result of dynamic cut object
res[[3]] # ggplot object of matplot of all clusters respectively.

##  The following examples takes too much times
# sample data
# fp <- system.file("extdata/nfpkm_rnsq.txt", package = "cornet")
# nfpkm <- read.table(fp, header=TRUE, stringsAsFactors = FALSE)

## dat1:
# dat1 <- nfpkm
# res1 <- cluster_mat(dat = dat1, distm = "spearman", clm = "average",
#     column = 5:ncol(dat1), method_dycut = "tree",
#     y_fctr = dat1$runs, x_fctr = dat1$days, rep_fctr = dat1$reps)

# dat2: no replicate and y_fctr
# dat2 <- nfpkm[which(nfpkm$runs==1)[seq(1,36, 3)] ,]
# dat2 <- data.frame(dat2[2], dat2[-1:-3][colSums(dat2[-1:-3]) != 0])

# res2 <- cluster_mat(dat = dat2, distm = "spearman", clm = "average",
#     column = 2:ncol(dat2), method_dycut = "tree", x_fctr=dat2$days)

## dat3: no y_fctr
# dat3 <- nfpkm[nfpkm$runs == 1, -1]
# dat3 <- data.frame(dat3[1:2], dat3[-1:-2][colSums(dat3[-1:-2]) != 0])
# res3 <- cluster_mat(dat = dat3, distm = "spearman", clm = "average",
#     column = 3:ncol(dat3), method_dycut = "tree", x_fctr=dat3$days, rep_fctr=dat3$reps)
}
