% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corgraph.R
\name{corgraph}
\alias{corgraph}
\title{Construction of a network graph from correlation matxix}
\usage{
corgraph(mat, thresh = seq(0.30,0.99,by=0.01))
}
\arguments{
\item{mat}{correlation matrix}

\item{thresh}{a vector of threshold sequence as iteration. The initial value as default, 'thresh =seq(0.30,0.99,by=0.01)'
if using an arbitral threshold, set a threshold value like as "thresh = 0.85".}
}
\value{
A list consists from one igraph object and two dataframes. The one is a weighted edge list, the another one is
    a result from ks-test of difference eigen sequence based on thresh correlation matrix.
}
\description{
Construction of a correlation network graph from correlation matrix through Random Matrix Theory-based methods.
   if negetive correlation exists, these information were added to result of igraph object.
}
\examples{
# # arguments
# nfpkm <- as.data.frame(rskodat::nfpkm)
# res_cld <- cornet::dycutdf(dat = nfpkm, column = -1:-4)$cluster_dat
# cormat <- cor(res_cld[["3"]])
# # create cor graph with a selected cluster
# res <- corgraph(mat = cormat)
# res[[1]]
# head(res[[2]])
# head(res[[3]])
}
