% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gethub.R
\name{gethub}
\alias{gethub}
\title{Node centrality}
\usage{
gethub(g, com_fun)
}
\arguments{
\item{g}{igraph object}

\item{com_fun}{character: function name of igraph for community detection}
}
\description{
get hub node.
}
\examples{
# sample data
data(cluster_dat)
cormat <- cor(cluster_dat[[2]])
res <- corgraph(mat = cormat)
g <- res[[1]]
gethub(g=g, com_fun="fastgreedy.community")
}
