% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igplot.R
\name{igplot}
\alias{igplot}
\title{simple plot of igraph object}
\usage{
igplot(ig, lay, v.l, v.l.c, v.f.c, v.s, v.c, e.c, e.w, ...)
}
\arguments{
\item{ig}{igraph object}

\item{lay}{graph layout}

\item{v.c, v.f.c, v.l, v.l.c, v.s}{vertex.color, vertex.frame.color, vertex.label, vertex.label.color, vertex.size}

\item{e.c, e.w}{edge.color, edge.width}

\item{...}{other arguments of plot.igraph}
}
\description{
several options has initial value, and
}
\examples{
# sample data
dat <- data.frame(
 S1 =c(43.26, 166.6, 12.53, 28.77, 114.7, 119.1, 118.9, 3.76, 32.73, 17.46),
 S2=c(40.89, 41.87, 39.55, 191.92, 79.7, 80.57, 156.69, 2.48, 11.99, 56.11),
 S3=c(5.05, 136.65, 42.09, 236.56, 99.76, 114.59, 186.95, 136.78, 118.8, 21.41)
 )
rownames(dat) <- paste0("G", 1:10)

# correlation matrix
cormat <- round(cor(t(dat)),2)

# threshold graph
res <- cornet::corgraph(mat=cormat)
g1 <- res[[1]]
igplot(ig=g1, v.s=15)

# complete graph
g2 <- cornet::matoedge(cormat)
 ewid <- abs(igraph::E(g2)$weight)
 ecol <-  ifelse(igraph::E(g2)$weight < 0 , "red", "grey80")
igplot(ig = g2, v.s = 15, e.c = ecol, e.w = ewid*4)

}
