% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matoedge.R
\name{matoedge}
\alias{matoedge}
\title{Convert a correlation matrix to a weighted edge list or igraph object}
\usage{
matoedge(mat, diag, zero.weight, format)
}
\arguments{
\item{mat}{A squared and synmetric matrix, like a correlation matirx.}

\item{diag}{default, FALSE}

\item{zero.weight}{default, FALSE}

\item{format}{select "igraph"(default) or "df" .}
}
\description{
Convert a correlation matrix to a weighted edge list from lower triangled data.
    if this matrix is named matrix, these names as node id.
}
\examples{
mat <- cor(iris[-5])
matoedge(mat = mat)
matoedge(mat = mat, diag = TRUE, format = "dataframe")
}
